fix: fixed email verification

+++ src/modules/Table/table.repository.ts
@@ -30,10 +30,10 @@ class TableRepository {
     return queryResult.rows[0]
   }
 
-  async deleteTableById(id: number): Promise<ITable> {
+  async deleteTableById(id: number, userId: number): Promise<ITable> {
     const queryResult: QueryResult<ITable> = await pool.query(
-      `DELETE FROM "Tables" WHERE id = $1 RETURNING id, "userId", name`,
-      [id],
+      `DELETE FROM "Tables" WHERE id = $1 AND "userId" = $2 RETURNING id, "userId", name`,
+      [id, userId],
     )
     if (!queryResult.rows.length) {
       throw new AppError(HttpStatusCode.NOT_FOUND, ErrorMessages.TABLE_NOT_FOUND)
@@ -41,10 +41,10 @@ class TableRepository {
     return queryResult.rows[0]
   }
 
-  async updateTable(id: number, name: string): Promise<ITable> {
+  async updateTable(id: number, name: string, userId: number): Promise<ITable> {
     const queryResult: QueryResult<ITable> = await pool.query(
-      `UPDATE "Tables" SET name = $2 WHERE id = $1 RETURNING id, "userId", name`,
-      [id, name],
+      `UPDATE "Tables" SET name = $2 WHERE id = $1 AND "userId" = $3 RETURNING id, "userId", name`,
+      [id, name, userId],
     )
     if (!queryResult.rows.length) {
       throw new AppError(HttpStatusCode.NOT_FOUND, ErrorMessages.TABLE_NOT_FOUND)
